requirements.txt
polars-1.38.1
duckdb-1.4.4
requests-2.32.5
streamlit-1.54.0
altair-6.0.0
plotly-6.5.2
